[refactor] working on message protocol
